refactor(frontend): split account settings components
